Numerous changes for Blog version 2.
Numerous changes for Blog version 2. The main change is that this is now
a configurable module. The module configuration screen is the first part
of the Blog install. It allows to set Blog's overall specs including
enabling an auto-publish/unpublish feature, commenting feature and a
blog style to install. See forum post for more info.